add hr task

+++ be/app/Repositories/Profile/ProfileEloquentRepository.php
@@ -39,41 +39,51 @@ class ProfileEloquentRepository extends EloquentRepository implements ProfileRep
     public function updateProfile(Request $request)
     {
         $profile = $this->_model->where('applier_id', $request->user()->id)->first();
-        //dd($profile->projects);
         $projectIds = Arr::pluck($profile->projects, 'id');
         $expDetailIds = Arr::pluck($profile->expDetail, 'id');
         $newProjects = $request->projects;
         $newExpDetail = $request->exp_detail;
+        $deleteProjects = null;
+        $deleteExpDetail = null;
 
-        $deleteProjects =
+        if ($projectIds) {
+            $deleteProjects =
             array_filter($projectIds, function ($id) use ($newProjects) {
                 return ! in_array($id, Arr::pluck($newProjects, 'id'));
             });
-
-        $deleteExpDetail =
+        }
+        if ($expDetailIds) {
+            $deleteExpDetail =
             array_filter($expDetailIds, function ($id) use ($newExpDetail) {
                 return ! in_array($id, Arr::pluck($newExpDetail, 'id'));
             });
-
-        $tempt = Arr::except($request->all(), ['projects', 'expDetail', 'workablePlaces', 'skills']);
+        }
+        $tempt = Arr::except($request->all(), ['projects', 'exp_detail', 'workable_places', 'skills']);
+        if (array_key_exists('gender', $tempt)) {
+            $tempt['gender'] += 2;
+        }
         $data = $profile->update($tempt);
-
         if ($data) {
-            if ($newExpDetail) {
+            /*if($newExpDetail) {
                 foreach ($newExpDetail as &$item) {
-                    $item['profile_id'] = $profile->id;
+                    $item["profile_id"] = $profile->id;
                 }
                 EXPdetail::query()->upsert($newExpDetail, ['id'], ['profile_id', 'place', 'content']);
             }
-            if ($newProjects) {
+            if($newProjects) {
                 foreach ($newProjects as &$item) {
-                    $item['profile_id'] = $profile->id;
+                    $item["profile_id"] = $profile->id;
                 }
                 Project::query()->upsert($newProjects, ['id'], ['profile_id', 'amount_of_member', 'start', 'end', 'technology', 'description']);
             }
 
-            EXPdetail::destroy($deleteExpDetail);
-            Project::destroy($deleteProjects);
+            if($deleteExpDetail) {
+                EXPdetail::destroy($deleteExpDetail);
+            }
+
+            if($deleteProjects) {
+                Project::destroy($deleteProjects);
+            }
 
             if ($request->workable_places) {
                 $profile->workablePlaces()->sync($request->workable_places);
@@ -81,8 +91,8 @@ class ProfileEloquentRepository extends EloquentRepository implements ProfileRep
             if ($request->skills) {
                 $profile->skills()->sync($request->skills);
             }
-
-            return $data;
+            */
+            return $profile;
         } else {
             return null;
         }
@@ -90,7 +100,7 @@ class ProfileEloquentRepository extends EloquentRepository implements ProfileRep
 
     public function createProfile(Request $request)
     {
-        $tempt = Arr::except($request->all(), ['projects', 'expDetail', 'workablePlaces', 'skills']);
+        $tempt = Arr::except($request->all(), ['projects', 'exp_detail', 'workable_places', 'skills']);
         $user = $request->user();
         //dd($user->id);
         if (! $user) {
